backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
sentence-transformers
torch
faiss-cpu
transformers
accelerate
bitsandbytes
tf-keras
requests